double to integer

+++ src/com/risk/model/Continent.java
@@ -10,10 +10,10 @@ import java.util.List;
 
 public class Continent  extends Land {
 double control;
-double db1;
+int db1;
 int playerId;
 
-public Continent(String new_name, double new_db1)
+public Continent(String new_name, int new_db1)
 {
 	super(new_name);
 	this.db1=new_db1;
@@ -22,7 +22,7 @@ public int GetContinentId()
 {
 	return this.id;
 	}
-public double GetDb1()
+public int GetDb1()
 {
 	return db1;
 	}
